Use 16 bits for the matching lengths in bpf files

+++ include/move_query.hpp
@@ -24,9 +24,9 @@ class MoveQuery {
         }
 
         void add_ml(uint64_t len_, bool to_stdout) {
-            uint32_t len = static_cast<uint32_t>(len_);
-            if (len_ > std::numeric_limits<uint32_t>::max()) {
-                len = std::numeric_limits<uint32_t>::max();
+            uint16_t len = static_cast<uint16_t>(len_);
+            if (len_ > std::numeric_limits<uint16_t>::max()) {
+                len = std::numeric_limits<uint16_t>::max();
             }
 
             matching_lens.push_back(len);
@@ -43,7 +43,7 @@ class MoveQuery {
         void add_cost(std::chrono::nanoseconds cost) { costs.push_back(cost); }
         void add_scan(uint64_t scan) { scans.push_back(scan); }
         void add_fastforward(uint64_t fastforward) { fastforwards.push_back(fastforward); }
-        std::vector<uint32_t>& get_matching_lengths() { return matching_lens; }
+        std::vector<uint16_t>& get_matching_lengths() { return matching_lens; }
         std::vector<uint64_t>& get_matching_colors() { return matching_colors; }
         std::vector<uint64_t>& get_sa_entries() { return sa_entries; }
         std::string& get_matching_lengths_string() { return matching_lengths_string; }
@@ -68,7 +68,7 @@ class MoveQuery {
         std::string query_string = "";
         std::string matching_lengths_string = "";
         std::vector<uint64_t> sa_entries;
-        std::vector<uint32_t> matching_lens;
+        std::vector<uint16_t> matching_lens;
         std::vector<uint64_t> matching_colors;
         std::vector<uint16_t> scans;
         std::vector<uint16_t> fastforwards;
